A Khora library compiles to WebAssembly and runs, with no imports
    $ KHORA_TARGET=wasm32-unknown-unknown khora build . --lib
    library src/lib.wasm from 8 module(s)

    const { instance } = await WebAssembly.instantiate(bytes, {});
    instance.exports.price(3n, 4n)        // 12n

`targets.md` steps two and three, for the one target that needs no sysroot.
`khora-rt` builds for `wasm32-unknown-unknown` — the dependencies that need an
operating system are behind a target predicate, and so are the seventeen
modules that use them — and `wasm-ld` links it, because a wasm module has no
libc to find and no CRT to link.

**Three bugs sat between "an object is emitted" and that, and each produced an
artefact that looked right.**

**The runtime ABI was the host's word size.** Seven runtime functions took a
Rust `usize` and were declared `i64` by the code generator: correct on the
three 64-bit targets, wrong on every other one. `KhoraHeader::refcount` was a
`usize` too, so the object header is 12 bytes on `wasm32` and 16 elsewhere
while the backend reads a `KHORA_HEADER_SIZE` compiled for the *host* — every
field offset out by four. `wasm-ld` was the first linker to say so, because it
checks signatures where an ELF or COFF linker matches on the name alone and
would have passed a 64-bit argument to a function expecting 32.

Fixed by making the ABI fixed-width rather than by teaching the backend the
pointer size. A contract between generated code and the runtime that changes
shape per target is one more thing to get wrong, and the other way round would
have had to be got right again at every call site. `khora-rt` takes `u64` and
narrows internally; the header is one layout everywhere, at four bytes an
object on a 32-bit target.

**`--allow-undefined` stopped the linker resolving.** It reads as "let the host
supply what is missing"; what it does is turn an unresolved symbol into an
import without looking further. `khora_overflow`, `khora_contain_enabled` and
`khora_export_call` all came out as `env.` imports *while defined in the
archive on the same command line*. The module linked, validated, exported
exactly the right two names, and could not have been instantiated.

**And the archive was the host's.** `runtime_archive` probes beside the `khora`
binary first, which is right for a native build and hands a wasm link an
x86-64 `.lib`. Silent under `--allow-undefined`; without it, `wasm-ld` says so
once per archive member.

**Size: 1.9 MB**, from naming each export rather than `--export-dynamic`, which
published 1,255 symbols from a three-function library and defeated dead-code
elimination. Most of what is left is Rust's formatting and backtrace machinery
reached from the trap handler, which a wasm trap needs neither of — the obvious
next cut. A Worker's free-plan limit is 1 MB compressed, so this is close
rather than comfortable, and the roadmap says so.

Two `std` files moved: `random.kh` became `random_native.kh`, because seeding a
PRNG needs a clock and a Worker has none, and seeding one from nothing is worse
than not having it.

The test asserts the **export and import sections**, since that is where two of
the three bugs showed: `price` and the control surface present, nothing
mangled, fewer than twenty in total, and no imports at all. It skips with a
message when the wasm runtime archive has not been built, because building it
downloads a target's standard library and a test that does that is a test
nobody runs.

Step four — the toolchain *fetching* a target's runtime — is now the only thing
between here and a `.wasm` on a fresh checkout.

`scripts/baseline.sh` clean end to end.

+++ crates/khora-rt/src/array.rs
@@ -56,7 +56,7 @@ fn array_word(array: *const u8, index: usize) -> usize {
 #[unsafe(no_mangle)]
 pub extern "C" fn khora_array_new(
     len: i64,
-    fill: usize,
+    fill: u64,
     stride: u8,
     boxed: u8,
     glue: Option<extern "C" fn(*mut u8)>,
@@ -83,7 +83,7 @@ pub extern "C" fn khora_array_new(
         fatal("an array that large cannot be allocated");
     };
 
-    let object = khora_alloc(fields, ARRAY_TAG);
+    let object = khora_alloc(fields as u64, ARRAY_TAG);
     // SAFETY: `khora_alloc` returned space for the header fields and the
     // elements, zeroed and aligned.
     unsafe {
